requirements.txt
Flask==2.2.5
line-bot-sdk==2.2.0
pika==1.3.1
